requirements.txt
pyaudio
SpeechRecognition
PyGame
regex
nltk
pyttsx3
googlesearch-python
beautifulsoup4